Cleanup: Move `debug-info` in its own sub-directory
