+++ plugins/lttng-utils/debug-info/bin-info.c
