+++ plugins/lttng-utils/debug-info/bin-info.h
