+++ plugins/lttng-utils/debug-info/crc32.c
