+++ plugins/lttng-utils/debug-info/crc32.h
