+++ plugins/lttng-utils/debug-info/debug-info.h
