+++ plugins/lttng-utils/debug-info/logging.c
