+++ plugins/lttng-utils/debug-info/logging.h
