+++ plugins/lttng-utils/debug-info/trace-ir-data-copy.h
