+++ plugins/lttng-utils/debug-info/trace-ir-mapping.h
