+++ plugins/lttng-utils/debug-info/trace-ir-metadata-copy.h
