+++ plugins/lttng-utils/debug-info/trace-ir-metadata-field-class-copy.h
