+++ plugins/lttng-utils/debug-info/utils.c
