+++ plugins/lttng-utils/debug-info/utils.h
