+++ tests/plugins/Makefile.am
@@ -16,7 +16,7 @@ endif # !ENABLE_BUILT_IN_PLUGINS
 
 if ENABLE_DEBUG_INFO
 test_dwarf_LDADD = \
-	$(top_builddir)/plugins/lttng-utils/libdebug-info.la \
+	$(top_builddir)/plugins/lttng-utils/debug-info/libdebug-info.la \
 	$(top_builddir)/logging/libbabeltrace-logging.la \
 	$(top_builddir)/common/libbabeltrace-common.la \
 	$(ELFUTILS_LIBS) \
@@ -24,7 +24,7 @@ test_dwarf_LDADD = \
 test_dwarf_SOURCES = test_dwarf.c
 
 test_bin_info_LDADD = \
-	$(top_builddir)/plugins/lttng-utils/libdebug-info.la \
+	$(top_builddir)/plugins/lttng-utils/debug-info/libdebug-info.la \
 	$(top_builddir)/logging/libbabeltrace-logging.la \
 	$(top_builddir)/common/libbabeltrace-common.la \
 	$(ELFUTILS_LIBS) \
