+++ tests/plugins/test_dwarf.c
@@ -25,7 +25,7 @@
 #include <stdint.h>
 #include <string.h>
 #include <unistd.h>
-#include <lttng-utils/dwarf.h>
+#include <lttng-utils/debug-info/dwarf.h>
 #include "tap/tap.h"
 
 #define NR_TESTS 15
